Chore: Refactor observer interface
